Renamed core folder to base
