+++ packages/desk-frame/src/app/MessageDialogOptions.ts
@@ -1,4 +1,4 @@
-import { ConfigOptions, LazyString, StringConvertible } from "../core/index.js";
+import { ConfigOptions, LazyString, StringConvertible } from "../base/index.js";
 
 /**
  * A representation of the contents of an alert or confirm dialog
