+++ packages/desk-frame/src/base/GlobalEmitter.ts
