+++ packages/desk-frame/src/base/LazyString.ts
