+++ packages/desk-frame/src/base/ManagedEvent.ts
