+++ packages/desk-frame/src/base/ManagedObject.ts
