+++ packages/desk-frame/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./core/index.js";
+export * from "./base/index.js";
 export * from "./app/index.js";
 export * from "./ui/index.js";
 export { JSX } from "./jsx/JSX.js";
