+++ packages/desk-frame/test/tests/base/bindings.ts
