+++ packages/desk-frame/test/tests/base/managedobject.ts
