+++ packages/desk-frame/test/tests/base/managedrecord.ts
